refactor url-sanitizer
