+++ index.js
@@ -27,11 +27,11 @@
 
 const assert = require('assert')
 const http = require('http')
-const fastDecode = require('fast-decode-uri-component')
 const isRegexSafe = require('safe-regex2')
 const { flattenNode, compressFlattenedNode, prettyPrintFlattenedNode, prettyPrintRoutesArray } = require('./lib/pretty-print')
 const Node = require('./node')
 const Constrainer = require('./lib/constrainer')
+const sanitizeUrl = require('./lib/url-sanitizer')
 
 const NODE_TYPES = Node.prototype.types
 const httpMethods = http.METHODS
@@ -369,10 +369,10 @@ Router.prototype.find = function find (method, path, derivedConstraints) {
     path = path.replace(FULL_PATH_REGEXP, '/')
   }
 
-  let sanitizedPath
+  let sanitizedUrl
   try {
-    sanitizedPath = sanitizeUrl(path)
-    path = sanitizedPath.path
+    sanitizedUrl = sanitizeUrl(path)
+    path = sanitizedUrl.path
   } catch (error) {
     return this.onBadUrl !== null
       ? this._onBadUrl(path)
@@ -486,10 +486,7 @@ Router.prototype.find = function find (method, path, derivedConstraints) {
       i = path.indexOf('/')
       if (i === -1) i = pathLen
       if (i > maxParamLength) return null
-      decoded = sanitizedPath.containsEncodedComponents
-        ? fastDecode(sanitizedPath.originPath.slice(idxInOriginalPath, idxInOriginalPath + i))
-        : originalPath.slice(idxInOriginalPath, idxInOriginalPath + i)
-
+      decoded = sanitizedUrl.sliceParameter(idxInOriginalPath, idxInOriginalPath + i)
       if (decoded === null) {
         return this.onBadUrl !== null
           ? this._onBadUrl(originalPath.slice(idxInOriginalPath, idxInOriginalPath + i))
@@ -504,10 +501,7 @@ Router.prototype.find = function find (method, path, derivedConstraints) {
 
     // wildcard route
     if (kind === NODE_TYPES.MATCH_ALL) {
-      decoded = sanitizedPath.containsEncodedComponents
-        ? fastDecode(sanitizedPath.originPath.slice(idxInOriginalPath))
-        : originalPath.slice(idxInOriginalPath)
-
+      decoded = sanitizedUrl.sliceParameter(idxInOriginalPath)
       if (decoded === null) {
         return this.onBadUrl !== null
           ? this._onBadUrl(originalPath.slice(idxInOriginalPath))
@@ -526,9 +520,7 @@ Router.prototype.find = function find (method, path, derivedConstraints) {
       i = path.indexOf('/')
       if (i === -1) i = pathLen
       if (i > maxParamLength) return null
-      decoded = sanitizedPath.containsEncodedComponents
-        ? fastDecode(sanitizedPath.originPath.slice(idxInOriginalPath, idxInOriginalPath + i))
-        : originalPath.slice(idxInOriginalPath, idxInOriginalPath + i)
+      decoded = sanitizedUrl.sliceParameter(idxInOriginalPath, idxInOriginalPath + i)
       if (decoded === null) {
         return this.onBadUrl !== null
           ? this._onBadUrl(originalPath.slice(idxInOriginalPath, idxInOriginalPath + i))
@@ -554,7 +546,7 @@ Router.prototype.find = function find (method, path, derivedConstraints) {
         while (i < pathLen && path.charCodeAt(i) !== 47 && path.charCodeAt(i) !== 45 && path.charCodeAt(i) !== 46) i++
         if (i > maxParamLength) return null
       }
-      decoded = fastDecode(originalPath.slice(idxInOriginalPath, idxInOriginalPath + i))
+      decoded = sanitizedUrl.sliceParameter(idxInOriginalPath, idxInOriginalPath + i)
       if (decoded === null) {
         return this.onBadUrl !== null
           ? this._onBadUrl(originalPath.slice(idxInOriginalPath, idxInOriginalPath + i))
@@ -571,12 +563,12 @@ Router.prototype.find = function find (method, path, derivedConstraints) {
   }
 }
 
-Router.prototype._getWildcardNode = function (node, path, len, derivedConstraints, params) {
+Router.prototype._getWildcardNode = function (node, sanitizedUrl, len, derivedConstraints, params) {
   if (node === null) return null
-  var decoded = fastDecode(path.slice(-len))
+  var decoded = sanitizedUrl.slice(-len)
   if (decoded === null) {
     return this.onBadUrl !== null
-      ? this._onBadUrl(path.slice(-len))
+      ? this._onBadUrl(sanitizedUrl.slice(-len))
       : null
   }
 
@@ -662,75 +654,6 @@ Router.prototype.all = function (path, handler, store) {
 }
 
 module.exports = Router
-
-// It must spot all the chars where decodeURIComponent(x) !== decodeURI(x)
-// The chars are: # $ & + , / : ; = ? @
-const uriComponentsCharMap = new Array(53).fill(0x00)
-uriComponentsCharMap[50] = new Array(103).fill(0x00)
-uriComponentsCharMap[50][51] = true // # '%23'
-uriComponentsCharMap[50][52] = true // $ '%24'
-uriComponentsCharMap[50][54] = true // & '%26'
-uriComponentsCharMap[50][66] = true // + '%2B'
-uriComponentsCharMap[50][98] = true // + '%2b'
-uriComponentsCharMap[50][67] = true // , '%2C'
-uriComponentsCharMap[50][99] = true // , '%2c'
-uriComponentsCharMap[50][70] = true // / '%2F'
-uriComponentsCharMap[50][102] = true // / '%2f'
-
-uriComponentsCharMap[51] = new Array(103).fill(0x00)
-uriComponentsCharMap[51][65] = true // : '%3A'
-uriComponentsCharMap[51][97] = true // : '%3a'
-uriComponentsCharMap[51][66] = true // ; '%3B'
-uriComponentsCharMap[51][98] = true // ; '%3b'
-uriComponentsCharMap[51][68] = true // = '%3D'
-uriComponentsCharMap[51][100] = true // = '%3d'
-uriComponentsCharMap[51][70] = true // ? '%3F'
-uriComponentsCharMap[51][102] = true // ? '%3f'
-
-uriComponentsCharMap[52] = new Array(49).fill(0x00)
-uriComponentsCharMap[52][48] = true // @ '%40'
-
-function sanitizeUrl (url) {
-  let originPath = url
-  let shouldDecode = false
-  let containsEncodedComponents = false
-  let highChar
-  let lowChar
-  for (var i = 0, len = url.length; i < len; i++) {
-    var charCode = url.charCodeAt(i)
-
-    if (shouldDecode && !containsEncodedComponents) {
-      if (highChar === 0 && uriComponentsCharMap[charCode]) {
-        highChar = charCode
-        lowChar = 0x00
-        continue
-      } else if (highChar && lowChar === 0 && uriComponentsCharMap[highChar][charCode]) {
-        containsEncodedComponents = true
-        continue
-      } else {
-        highChar = undefined
-        lowChar = undefined
-      }
-    }
-
-    // Some systems do not follow RFC and separate the path and query
-    // string with a `;` character (code 59), e.g. `/foo;jsessionid=123456`.
-    // Thus, we need to split on `;` as well as `?` and `#`.
-    if (charCode === 63 || charCode === 59 || charCode === 35) {
-      originPath = url.slice(0, i)
-      break
-    } else if (charCode === 37) {
-      shouldDecode = true
-      highChar = 0x00
-    }
-  }
-  const decoded = shouldDecode ? decodeURI(originPath) : originPath
-  return {
-    path: decoded,
-    originPath,
-    containsEncodedComponents
-  }
-}
 
 function getClosingParenthensePosition (path, idx) {
   // `path.indexOf()` will always return the first position of the closing parenthese,
